feat: move InterstitialScreenView and ViewModule to storybook assets
