--- packages/ibm-products/src/components/InterstitialScreenView/InterstitialScreenView.mdx
@@ -1,48 +1,0 @@
-import { Story, Controls, Source, Canvas } from '@storybook/addon-docs';
-import { CodesandboxLink } from '../../global/js/utils/story-helper';
-import { InterstitialScreenView } from '.';
-import * as stories from './InterstitialScreenView.stories';
-
-# InterstitialScreenView
-
-[InterstitialScreen usage guidelines](https://pages.github.ibm.com/carbon/ibm-products/components/onboarding/interstitial-screen/)
-
-## Table of Contents
-
-- [Overview](#overview)
-- [About Onboarding](#about-onboarding)
-- [Example usage](#example-usage)
-- [Component API](#component-api)
-
-## Overview
-
-InterstitialScreenView is designed as a child of **InterstitialScreen**. The
-provided `stepTitle` prop is used as an aria-label within the component, as well
-as by the parent InterstitialScreen component when multiple
-InterstitialScreenView's are used in a Carbon
-[ProgressIndicator](https://carbondesignsystem.com/components/progress-indicator/usage/)
-to display the user's progression.
-
-## About Onboarding
-
-[Onboarding](https://pages.github.ibm.com/security/security-design/teams-across-security/onboarding/overview/)
-is a continuous learning methodology and framework that aims to orient, onboard,
-explain, educate, and cultivate novice users into high-functioning power users.
-
-## Example usage
-
-<Canvas>
-  <Story of={stories.interstitialScreenView} />
-</Canvas>
-
-## Code sample
-
-```jsx
-<InterstitialScreenView stepTitle="Step 1">
-  <div>Hello world</div>
-</InterstitialScreenView>
-```
-
-## Component API
-
-<Controls />
