--- packages/ibm-products/src/components/InterstitialScreenViewModule/InterstitialScreenViewModule.mdx
@@ -1,46 +1,0 @@
-import { Story, Controls, Source, Canvas } from '@storybook/addon-docs';
-import { CodesandboxLink } from '../../global/js/utils/story-helper';
-import { InterstitialScreenViewModule } from '.';
-import * as stories from './InterstitialScreenViewModule.stories';
-
-# InterstitialScreenViewModule
-
-[InterstitialScreen usage guidelines](https://pages.github.ibm.com/carbon/ibm-products/components/onboarding/interstitial-screen/)
-
-## Table of Contents
-
-- [Overview](#overview)
-- [About Onboarding](#about-onboarding)
-- [Example usage](#example-usage)
-- [Component API](#component-api)
-
-## Overview
-
-InterstitialScreenViewModule is an optional view module to jump start
-development and is designed to be used as a child of InterstitialScreenView.
-
-## About Onboarding
-
-[Onboarding](https://pages.github.ibm.com/security/security-design/teams-across-security/onboarding/overview/)
-is a continuous learning methodology and framework that aims to orient, onboard,
-explain, educate, and cultivate novice users into high-functioning power users.
-
-## Example usage
-
-<Canvas>
-  <Story of={stories.interstitialScreenViewModule} />
-</Canvas>
-
-## Code sample
-
-```jsx
-<InterstitialScreenViewModule
-  className="MyInterstitialScreenViewModule"
-  description="This is an example of a description"
-  title="Example title"
-/>
-```
-
-## Component API
-
-<Controls />
