--- packages/ibm-products/src/components/InterstitialScreenViewModule/_storybook-styles.scss
@@ -1,8 +1,0 @@
-//
-// Copyright IBM Corp. 2024, 2024
-//
-// This source code is licensed under the Apache-2.0 license found in the
-// LICENSE file in the root directory of this source tree.
-//
-
-// TODO: add any additional styles used by InterstitialScreenViewModule.stories.js
