--- packages/ibm-products/src/components/InterstitialScreenViewModule/index.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright IBM Corp. 2024, 2024
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export { InterstitialScreenViewModule } from './InterstitialScreenViewModule';
-export type { InterstitialScreenViewModuleProps } from './InterstitialScreenViewModule';
